Ready the Products window

+++ Presentation/Forms/ProductsWindow.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using Domain.Processors;
-using Microsoft.IdentityModel.Tokens;
 using SupportLayer.Models;
 using System.Collections.ObjectModel;
 using System.Linq;
@@ -15,13 +14,14 @@ public partial class ProductsWindow : Window, ISpeciesRequestor
     private ObservableCollection<Species> _species;
     private SpeciesProcessor _speciesProcessor;
     private ProductProcessor _productProcessor;
+    private Product _productModel;
 
     public ProductsWindow()
     {
         InitializeComponent();
-        _species = new ObservableCollection<Species>();
         _speciesProcessor = new SpeciesProcessor();
         _productProcessor = new ProductProcessor();
+        _productModel = new Product();
         LoadData();
     }
 
@@ -53,6 +53,11 @@ public partial class ProductsWindow : Window, ISpeciesRequestor
             AddEditSpeciesWindow window = new AddEditSpeciesWindow(species);
             window.ShowDialog();
             dgSpecies.Items.Refresh();
+        }
+        else
+        {
+            MessageBox.Show("Debe seleccionar el registro que desea editar."
+                , "", MessageBoxButton.OK, MessageBoxImage.Information);
         }
     }
 
@@ -93,52 +98,94 @@ public partial class ProductsWindow : Window, ISpeciesRequestor
     {
         if (dgSpecies.SelectedItem is Species species)
         {
-            lstVarieties.ItemsSource = species.Products.OrderBy(x => x.Variety);            
+            lstVarieties.ItemsSource = species.Products.OrderBy(x => x.Variety);
+            btnRemoveVariety.IsEnabled = true;
+            txtNewVariety.Text = "";
         }
     }
 
     private void btnAddVariety_Click(object sender, RoutedEventArgs e)
     {
         //LATER - Create an string extension method in the SupportLayer project to check if a string is null or whitespace(this is whitespace"   ")
-        if (txtNewVariety.Text != string.Empty && dgSpecies.SelectedItem is Species species)
+
+        ValidateDataType();
+
+        if (_productModel.Id == 0)
         {
-            Product newProduct = new Product();
-            newProduct.SpecieId = species.Id;
-            newProduct.Variety = txtNewVariety.Text;
+            ((Species)dgSpecies.SelectedItem).Products.Add(_productModel);
+        }
 
-            _productProcessor.SaveProduct(newProduct);
-
-            species.Products.Add(newProduct);
-
+        if (_productProcessor.SaveProduct(_productModel) == true)
+        {
             RefreshListBox();
 
+            _productModel = new Product();
+
             txtNewVariety.Text = "";
+            btnAddVariety.Content = "Agregar";
+            btnRemoveVariety.IsEnabled = true;
         }
+        else
+        {
+            ShowError();
+        }
+    }
+
+    private void ValidateDataType()
+    {
+        _productModel.Variety = txtNewVariety.Text;
+        _productModel.SpecieId = ((Species)dgSpecies.SelectedItem).Id;
+    }
+
+    private void ShowError()
+    {
+        MessageBox.Show(_productProcessor.Error);
     }
 
     private void RefreshListBox()
     {
-            Species species = (Species)dgSpecies.SelectedItem;            
-            lstVarieties.ItemsSource = null;
-            lstVarieties.ItemsSource = species.Products.OrderBy(x => x.Variety);
+        Species species = (Species)dgSpecies.SelectedItem;
+        lstVarieties.ItemsSource = null;
+        lstVarieties.ItemsSource = species.Products.OrderBy(x => x.Variety);
     }
 
     private void btnRemoveVariety_Click(object sender, RoutedEventArgs e)
     {
-            Species species = (Species)dgSpecies.SelectedItem;        
-            Product product = (Product)lstVarieties.SelectedItem;
-        
-            _productProcessor.DeleteProduct(product.Id);
+        if (lstVarieties.SelectedItem is Product product)
+        {
+            if (MessageBox.Show("Esta seguro que desea eliminar este registro?"
+                , "Eliminar registro"
+                , MessageBoxButton.YesNo
+                , MessageBoxImage.Warning) == MessageBoxResult.Yes)
+            {
+                _productProcessor.DeleteProduct(product.Id);
 
-            species.Products.Remove(product);
+                ((Species)dgSpecies.SelectedItem).Products.Remove(product);
 
-            RefreshListBox();      
+                RefreshListBox();
+            }
+        }
+        else
+        {
+            MessageBox.Show("Debe seleccionar el registro que desea eliminar."
+                , "", MessageBoxButton.OK, MessageBoxImage.Information);
+        }
     }
 
     private void lstVarieties_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
     {
-
+        //LATER - In the database and all the logic above change the table Product by Variety
+        if (lstVarieties.SelectedItem is Product product)
+        {
+            _productModel = product;
+            txtNewVariety.Text = product.Variety;
+            btnAddVariety.Content = "Editar";
+            btnRemoveVariety.IsEnabled = false;
+        }
+        else
+        {
+            MessageBox.Show("Debe seleccionar el registro que desea editar."
+                , "", MessageBoxButton.OK, MessageBoxImage.Information);
+        }
     }
-
-
 }
